chore: merge ui progress
