--- src/app/pages/send/ordinal-inscription/types.ts
@@ -1,3 +1,0 @@
-interface Values {
-  recipient: string;
-}
