+++ src/shared/route-urls.ts
@@ -64,7 +64,6 @@ export enum RouteUrls {
 
   // Send ordinal inscriptions
   SendOrdinalInscription = '/send/ordinal-inscription',
-  SendOrdinalInscriptionSend = '/send/ordinal-inscription/send',
   SendOrdinalInscriptionChooseFee = '/send/ordinal-inscription/choose-fee',
   SendOrdinalInscriptionReview = '/send/ordinal-inscription/review',
   SendOrdinalInscriptionSent = '/send/ordinal-inscription/sent',
